requirements.txt
# Deep Learning & ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
facenet-pytorch>=2.5.3
dlib>=19.24.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Framework
streamlit>=1.28.0
flask>=2.3.0

# PDF Generation
reportlab>=4.0.0
fpdf2>=2.7.0
pillow>=10.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
werkzeug>=2.3.0

# Image Processing
albumentations>=1.3.0
timm>=0.9.0

# Logging & Monitoring
loguru>=0.7.0

# File Handling
python-magic-bin>=0.4.14; sys_platform == 'win32'
